Try templating the modules out better
There's a lot of copy/paste between the modules and while refactoring
things/experimenting with new architectures, I thought it would be
helpful to template them out more explicitly. Unfortunately that broke
everything and I'm not even convinced that the complexity flake.parts
brings is worth it in the end.